unreliable no longer supports Rails 5.0 and 5.1
For now, at least, unreliable's addition of an order triggers a bug that
I believe is this issue:

https://github.com/rails/rails/issues/13496

In Rails 5.0 and 5.1, for sqlite and postgres, this can generate buggy
update_all statements that can update every row in the table instead of
the one row desired! This issue appears to have been resolved in 5.2,
or at least unreliable's tests pass there. I'll continue to investigate.